Removed version from request packages

+++ src/main/java/de/fernunihagen/dna/scalephant/network/server/ClientConnectionHandler.java
@@ -637,7 +637,7 @@ public class ClientConnectionHandler implements Runnable {
 		final ByteBuffer packageHeader = readNextPackageHeader();
 
 		final short packageSequence = NetworkPackageDecoder.getRequestIDFromRequestPackage(packageHeader);
-		final byte packageType = NetworkPackageDecoder.getPackageTypeFromRequest(packageHeader);
+		final short packageType = NetworkPackageDecoder.getPackageTypeFromRequest(packageHeader);
 		
 		boolean readFurtherPackages = true;
